Update documentation and source code for pre-update profile action
