--- pre-update-user-profile-action/vercel/nodejs/api/index.js
@@ -1,3 +1,0 @@
-module.exports = (req, res) => {
-    res.status(200).send('Hello from the home endpoint!');
-};
