--- pre-update-user-profile-action/vercel/nodejs/vercel.json
@@ -1,1 +1,0 @@
-{ "version": 2, "rewrites": [{ "source": "/(.*)", "destination": "/api" }] }
