+++ pre-update-user-profile-samples/pre-profile-update-validation-service-node-vercel/api/index.js
@@ -1,0 +1,3 @@
+module.exports = (req, res) => {
+    res.status(200).json({"status": "ok", "message": "Service is running."});
+};
