+++ pre-update-user-profile-samples/pre-profile-update-validation-service-node-vercel/vercel.json
@@ -1,0 +1,1 @@
+{ "version": 2, "rewrites": [{ "source": "/(.*)", "destination": "/api" }] }
